+++ Source/TencentVoiceForUnreal/Private/VoiceClient.cpp
@@ -88,7 +88,7 @@ int32 UVoiceClient::SetAppInfo(const FString& appID, const FString& appKey, cons
 
 	if (nullptr != m_voiceengine)
 	{
-		ErrorCode = static_cast<int32>(m_voiceengine->SetAppInfo(TCHAR_TO_ANSI(*appID), TCHAR_TO_ANSI(*appKey), TCHAR_TO_ANSI(*OpenID)));
+		ErrorCode = m_voiceengine->SetAppInfo(TCHAR_TO_ANSI(*appID), TCHAR_TO_ANSI(*appKey), TCHAR_TO_ANSI(*OpenID));
 		return ErrorCode;
 	}
 
@@ -98,7 +98,7 @@ int32 UVoiceClient::SetAppInfo(const FString& appID, const FString& appKey, cons
 
 int32 UVoiceClient::InitVoiceEngine()
 {
-	return static_cast<int32>(m_voiceengine->Init());
+	return m_voiceengine->Init();
 }
 
 int32 UVoiceClient::SetNotify(UNotifyBase* NotifyInstance)
@@ -107,7 +107,7 @@ int32 UVoiceClient::SetNotify(UNotifyBase* NotifyInstance)
 
 	if (nullptr != NotifyInstance)
 	{
-		ErrorCode = static_cast<int32>(m_voiceengine->SetNotify(NotifyInstance));
+		ErrorCode = m_voiceengine->SetNotify(NotifyInstance);
 		return ErrorCode;
 	}
 	ErrorCode = -1;
@@ -116,12 +116,12 @@ int32 UVoiceClient::SetNotify(UNotifyBase* NotifyInstance)
 
 int32 UVoiceClient::SetMode(EVoiceMode VoiceMode)
 {
-	return static_cast<int32>(m_voiceengine->SetMode(static_cast<gcloud_voice::IGCloudVoiceEngine::GCloudVoiceMode>(VoiceMode)));
+	return m_voiceengine->SetMode(static_cast<gcloud_voice::IGCloudVoiceEngine::GCloudVoiceMode>(VoiceMode));
 }
 
 int32 UVoiceClient::EnableMultiRoom(bool bEnable)
 {
-	return static_cast<int32>(m_voiceengine->EnableMultiRoom(bEnable));
+	return m_voiceengine->EnableMultiRoom(bEnable);
 }
 
 int32 UVoiceClient::JoinTeamRoom(const FString & RoomName, int32 msTimeout)
@@ -133,7 +133,7 @@ int32 UVoiceClient::JoinTeamRoom(const FString & RoomName, int32 msTimeout)
 
 	if (!JoinedRoomName.Contains(RoomName))
 	{
-		return static_cast<int32>(m_voiceengine->JoinTeamRoom(TCHAR_TO_ANSI(*RoomName), msTimeout));
+		return m_voiceengine->JoinTeamRoom(TCHAR_TO_ANSI(*RoomName), msTimeout);
 	}
 	return -1;
 }
@@ -147,7 +147,7 @@ int32 UVoiceClient::JoinNationalRoom(const FString & RoomName, EVoiceMemberRole 
 
 	if (!JoinedRoomName.Contains(RoomName))
 	{
-		return static_cast<int32>(m_voiceengine->JoinNationalRoom(TCHAR_TO_ANSI(*RoomName), static_cast<gcloud_voice::IGCloudVoiceEngine::GCloudVoiceMemberRole>(MemberRole), msTimeout));
+		return m_voiceengine->JoinNationalRoom(TCHAR_TO_ANSI(*RoomName), static_cast<gcloud_voice::IGCloudVoiceEngine::GCloudVoiceMemberRole>(MemberRole), msTimeout);
 	}
 	return -1;
 }
@@ -156,62 +156,62 @@ int32 UVoiceClient::QuitRoom(const FString & RoomName, int32 msTimeout)
 {
 	if (JoinedRoomName.Contains(RoomName))
 	{
-		return static_cast<int32>(m_voiceengine->QuitRoom(TCHAR_TO_ANSI(*RoomName), msTimeout));
+		return m_voiceengine->QuitRoom(TCHAR_TO_ANSI(*RoomName), msTimeout);
 	}
 	return -1;
 }
 
 int32 UVoiceClient::SetServerInfo(const FString & ServerAddr)
 {
-	return static_cast<int32>(m_voiceengine->SetServerInfo(TCHAR_TO_ANSI(*ServerAddr)));
+	return m_voiceengine->SetServerInfo(TCHAR_TO_ANSI(*ServerAddr));
 }
 
 int32 UVoiceClient::OnPause()
 {
-	return static_cast<int32>(m_voiceengine->Pause());
+	return m_voiceengine->Pause();
 }
 
 int32 UVoiceClient::OnResume()
 {
-	return static_cast<int32>(m_voiceengine->Resume());
+	return m_voiceengine->Resume();
 }
 
 int32 UVoiceClient::TestMic()
 {
-	return static_cast<int32>(m_voiceengine->TestMic());
+	return m_voiceengine->TestMic();
 }
 
 int32 UVoiceClient::GetMicLevel()
 {
-	return static_cast<int32>(m_voiceengine->GetMicLevel(true));
+	return m_voiceengine->GetMicLevel(true);
 }
 
 int32 UVoiceClient::OpenMic()
 {
-	return static_cast<int32>(m_voiceengine->OpenMic());
+	return m_voiceengine->OpenMic();
 }
 
 int32 UVoiceClient::CloseMic()
 {
-	return static_cast<int32>(m_voiceengine->CloseMic());
+	return m_voiceengine->CloseMic();
 }
 
 int32 UVoiceClient::OpenSpeaker()
 {
-	return static_cast<int32>(m_voiceengine->OpenSpeaker());
+	return m_voiceengine->OpenSpeaker();
 }
 
 int32 UVoiceClient::CloseSpeaker()
 {
-	return static_cast<int32>(m_voiceengine->CloseSpeaker());
+	return m_voiceengine->CloseSpeaker();
 }
 
 int32 UVoiceClient::SetMicVolume(int vol)
 {
-	return static_cast<int32>(m_voiceengine->SetMicVolume(vol));
+	return m_voiceengine->SetMicVolume(vol);
 }
 
 int32 UVoiceClient::SetSpeakerVolume(int vol)
 {
-	return static_cast<int32>(m_voiceengine->SetSpeakerVolume(vol));
+	return m_voiceengine->SetSpeakerVolume(vol);
 }
